# doc/OpenAPI.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "K-XYZ 智能餐盒全链路 API",
    "version": "4.4.0",
    "description": "根据 D:\\MIT\\newdev\\doc\\API.md（v4.4版）整理生成的 OpenAPI 描述。除健康检查、硬件遥测上报、注册、登录外，其余接口默认需要 Bearer JWT。"
  },
  "servers": [
    {
      "url": "https://api.mit.chenyuxia.com/api/v1",
      "description": "生产环境"
    },
    {
      "url": "http://127.0.0.1:8080/api/v1",
      "description": "本地开发"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/users/me/profile": {
      "get": {
        "tags": [
          "User Profile"
        ],
        "summary": "获取个人健康画像",
        "operationId": "getUserProfile",
        "responses": {
          "200": {
            "description": "画像查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                },
                "example": {
                  "user_id": "string",
                  "gender": "male",
                  "age": 22,
                  "height_cm": 178.5,
                  "weight_kg": 75.0,
                  "updated_at": "2026-03-27T10:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "用户尚未填写健康画像",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileNotFoundResponse"
                },
                "example": {
                  "error": "profile_not_found",
                  "message": "用户尚未填写健康画像"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "User Profile"
        ],
        "summary": "更新或保存个人健康画像",
        "description": "后端需执行 Upsert 逻辑：记录存在则更新，不存在则新建。",
        "operationId": "upsertUserProfile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertUserProfileRequest"
              },
              "example": {
                "gender": "male",
                "age": 22,
                "height_cm": 178.5,
                "weight_kg": 75.0
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "画像保存成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpsertUserProfileResponse"
                },
                "example": {
                  "message": "画像保存成功",
                  "updated_at": "2026-03-27T10:05:00Z"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "除 /ping、/hardware/telemetry、/auth/register、/auth/login 外，其余接口默认需要 Bearer Token。"
      }
    },
    "schemas": {
...
      "Gender": {
        "type": "string",
        "enum": [
          "male",
          "female",
          "other"
        ]
      },
      "UserProfileResponse": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "age": {
            "type": "integer"
          },
          "height_cm": {
            "type": "number"
          },
          "weight_kg": {
            "type": "number"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "user_id",
          "gender",
          "age",
          "height_cm",
          "weight_kg",
          "updated_at"
        ]
      },
      "ProfileNotFoundResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "const": "profile_not_found"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ]
      },
      "UpsertUserProfileRequest": {
        "type": "object",
        "properties": {
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "age": {
            "type": "integer"
          },
          "height_cm": {
            "type": "number"
          },
          "weight_kg": {
            "type": "number"
          }
        },
        "required": [
          "gender",
          "age",
          "height_cm",
          "weight_kg"
        ]
      },
      "UpsertUserProfileResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "message",
          "updated_at"
        ]
      }
    }
  }
}